start c++ implementation

+++ wasm/include/libc/stdint.h
@@ -6,7 +6,7 @@ typedef unsigned long long uint64_t;
 typedef unsigned int uint32_t;
 typedef unsigned short uint16_t;
 typedef unsigned char uint8_t;
-typedef uint32_t size_t;
+typedef unsigned long size_t;
 typedef uint32_t uintptr_t;
 
 #define NULL 0
